requirements.txt
Flask==2.0.1
aiohttp==3.7.4
python-dotenv==0.19.0